requirements.txt
anthropic>=0.40.0
flask>=3.0.0
gunicorn>=23.0.0
python-dotenv>=1.0.0
jsonschema>=4.23.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.150.0
playwright>=1.48.0
pillow>=11.0.0
